feat: handle logout socket message
