+++ frontend/src/handlers/game/matchMakingHandlers.js
@@ -7,6 +7,7 @@ import {
 import getRouter from "../../core/router.js";
 import GameInviteModal from "../../components/Lobby/GameInviteModal.js";
 import fetchUserInfo from "../../api/user/fetchUserInfo.js";
+import logoutHandler from "../auth/logoutHandler.js";
 
 const matchFindHandler = (matchType) => {
   const { sendSocket } = gameSocket();
@@ -183,6 +184,15 @@ const receiveGameInviteHandler = (removeObservers) => {
   removeObservers.push(removeObserver);
 };
 
+const receiveLogoutHandler = (removeObservers) => {
+  const { addSocketObserver } = gameSocket();
+
+  const removeObserver = addSocketObserver("user_logout", () => {
+    logoutHandler();
+  });
+  removeObservers.push(removeObserver);
+};
+
 export {
   matchFindHandler,
   cancleMatchFindHandler,
@@ -196,4 +206,5 @@ export {
   getGameInfoHandler,
   tournamentBeginHandler,
   receiveGameInviteHandler,
+  receiveLogoutHandler,
 };
